+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'movilidad-v1.11.1'
+const CACHE_NAME = 'movilidad-v1.11.2'
 
 self.addEventListener('install', (event) => {
   event.waitUntil(